refactor: setting 페이지 관련 layout 마크업 구조 개선 #12
