+++ src/constants/page.ts
@@ -8,6 +8,11 @@ export const PAGE_PATH = {
   licenses: '/setting/licenses',
 };
 
+export const PAGE_TITLE_MAPPER: Record<string, string> = {
+  '/setting': '설정',
+  '/setting/licenses': '오픈소스 라이선스 이용고지',
+};
+
 export const QUERY_STRING = {
   search: 'search',
 };
